+++ tmc_wrs_gazebo_launch/CHANGELOG.rst
@@ -6,3 +6,10 @@ Changelog for package tmc_wrs_gazebo_launch
 -------------------
 * Initial commit
 * Contributors: Nobuyuki Matsuno, Takashi Yamamoto, Yasukata Yokochi, Yuka Hashiguchi
+
+2.0.0 (2025-07-10)
+-------------------
+* Support ROS2 Humble
+* Change the build type to amant-python
+* Contributors: Tomoaki Fujino (Hibikino-Musashi@Home)
+